# lib/api-zod/src/generated/types/recallDatasetCoverage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecallDatasetCoverage {
  datasetId: string;
  manufacturer: string;
  label: string;
  sourceFile: string;
  sourceDescription: string;
  coveragePeriod: string;
  synthetic: boolean;
  recordCount: number;
  estimatedUnitsAffected: number;
}