# lib/api-zod/src/generated/types/recallDatasetPage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecallDatasetCoverage } from './recallDatasetCoverage';
import type { RecallRecord } from './recallRecord';

export interface RecallDatasetPage {
  items: RecallRecord[];
  page: number;
  pageSize: number;
  total: number;
  totalPages: number;
  coverage: RecallDatasetCoverage;
}