# lib/api-zod/src/generated/types/recallRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecallRecord {
  campaignId: string;
  vehicleModels: string;
  component: string;
  errorDescription: string;
  estimatedUnitsAffected: number;
  remedyAction: string;
}